+++ internal/controller/dispatcher.go
@@ -80,139 +80,6 @@ func (rd *ResponseDispatcher) Run(ctx context.Context) {
 
 }
 
-type PayloadHandler struct {
-	Account    string
-	NodeID     string
-	Writer     *kafka.Writer
-	ReceptorSM *ReceptorStateMachine
-}
-
-func (rd *PayloadHandler) GetKey() string {
-	return fmt.Sprintf("%s:%s", rd.Account, rd.NodeID)
-}
-
-func (rd PayloadHandler) HandleMessage(ctx context.Context, m protocol.Message /*, receptorID string*/) error {
-	type ResponseMessage struct {
-		Account      string      `json:"account"`
-		Sender       string      `json:"sender"`
-		MessageType  string      `json:"message_type"`
-		MessageID    string      `json:"message_id"`
-		Payload      interface{} `json:"payload"`
-		Code         int         `json:"code"`
-		InResponseTo string      `json:"in_response_to"`
-		Serial       int         `json:"serial"`
-	}
-
-	if m.Type() != protocol.PayloadMessageType {
-		log.Printf("Unable to dispatch message (type: %d): %s", m.Type(), m)
-		return nil
-	}
-
-	payloadMessage, ok := m.(*protocol.PayloadMessage)
-	if !ok {
-		log.Println("Unable to convert message into PayloadMessage")
-		return nil
-	}
-
-	// FIXME:
-	/*
-		// verify this message was meant for this receptor/peer (probably want a uuid here)
-		if payloadMessage.RoutingInfo.Recipient != receptorID {
-			log.Println("Recieved message that was not intended for this node")
-			return nil
-		}
-	*/
-
-	responseMessage := ResponseMessage{
-		Account:      rd.Account,
-		Sender:       payloadMessage.RoutingInfo.Sender,
-		MessageID:    payloadMessage.Data.MessageID,
-		MessageType:  payloadMessage.Data.MessageType,
-		Payload:      payloadMessage.Data.RawPayload,
-		Code:         payloadMessage.Data.Code,
-		InResponseTo: payloadMessage.Data.InResponseTo,
-		Serial:       payloadMessage.Data.Serial,
-	}
-
-	log.Printf("Dispatching response:%+v", responseMessage)
-
-	jsonResponseMessage, err := json.Marshal(responseMessage)
-	if err != nil {
-		log.Println("JSON marshal of ResponseMessage failed, err:", err)
-		return nil
-	}
-
-	rd.Writer.WriteMessages(ctx,
-		kafka.Message{
-			Key:   []byte(payloadMessage.Data.InResponseTo),
-			Value: jsonResponseMessage,
-		})
-
-	return nil
-}
-
-type RouteTableHandler struct {
-	ReceptorSM *ReceptorStateMachine
-}
-
-func (rth RouteTableHandler) HandleMessage(ctx context.Context, m protocol.Message) error {
-
-	log.Printf("inside RouteTableHandler...statemachine:%+v", rth.ReceptorSM)
-
-	if rth.ReceptorSM.handshakeComplete == false {
-		log.Println("Received ROUTE message before handshake was complete")
-		// FIXME:  send an error on the ErrorChannel and shutdown connection?!?!
-		return nil
-	}
-
-	if m.Type() != protocol.RouteTableMessageType {
-		log.Printf("Invalid message type (type: %d): %v", m.Type(), m)
-		return nil
-	}
-
-	routingTableMessage, ok := m.(*protocol.RouteTableMessage)
-	if !ok {
-		log.Println("Unable to convert message into RouteTableMessage")
-		return nil
-	}
-
-	log.Printf("**** got routing table message!!  %+v", routingTableMessage)
-
-	rth.ReceptorSM.routingTableReceived = true
-
-	return nil
-}
-
-type HandshakeHandler struct {
-	ControlChannel chan protocol.Message
-	ErrorChannel   chan error
-	ReceptorSM     *ReceptorStateMachine
-}
-
-func (hh HandshakeHandler) HandleMessage(ctx context.Context, m protocol.Message) error {
-	if m.Type() != protocol.HiMessageType {
-		hh.ErrorChannel <- fmt.Errorf("Invalid message type (type: %d): %v", m.Type(), m)
-		return nil
-	}
-
-	hiMessage, ok := m.(*protocol.HiMessage)
-	if !ok {
-		hh.ErrorChannel <- fmt.Errorf("Unable to convert message into HiMessage")
-		return nil
-	}
-
-	log.Printf("**** got hi message!!  %+v", hiMessage)
-
-	// FIXME: pass the read node name over to the client
-	responseHiMessage := protocol.HiMessage{Command: "HI", ID: "c.config.ReceptorControllerNodeId"}
-
-	hh.ControlChannel <- &responseHiMessage // FIXME:  Why a pointer here??
-
-	hh.ReceptorSM.handshakeComplete = true
-
-	return nil
-}
-
 type MessageDispatcherFactory struct {
 	readerConfig *queue.ConsumerConfig
 }
