+++ internal/controller/receptor.go
@@ -1,9 +1,21 @@
 package controller
 
-type Receptor struct {
+type MeshConnection struct {
+	peerNodeID string
 }
 
 type ReceptorStateMachine struct {
 	handshakeComplete    bool
 	routingTableReceived bool
+
+	account    string
+	peerNodeID string
+
+	connections map[string]MeshConnection
+
+	capabilities interface{}
+	/*
+	   edges
+	   seen
+	*/
 }
